+++ app/Services/XNova.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Services;
+
+class XNova
+{
+    const VERSION = '0.5.1-dev';
+}