+++ resources/lang/en/admin/common.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'no-access' => 'You do not have access to this page',
+];