+++ resources/lang/en/admin/errors.php
@@ -1,0 +1,11 @@
+<?php
+
+return [
+    'clear' => 'clear the list',
+    'count' => 'error(s) total',
+    'date' => 'Date',
+    'list-title' => 'Error messages',
+    'player' => 'Player ID',
+    'title' => 'Error reports',
+    'type' => 'Type',
+];