+++ resources/lang/en/app.php
@@ -2,6 +2,7 @@
 
 return [
     'error' => 'Error',
+    'format-datetime' => 'd/m/Y h:i:s',
     'name' => 'XNova',
     'success' => 'Success',
 ];